+++ src/apple/fastlane.ts
@@ -77,7 +77,6 @@ function addSentryToLane(
   lane: { index: number; length: number; name: string },
   org: string,
   project: string,
-  token: string,
 ): string {
   const laneContent = content.slice(lane.index, lane.index + lane.length);
   const sentryCLIMatch = /sentry_cli\s*\([^)]+\)/gim.exec(laneContent);
@@ -85,7 +84,7 @@ function addSentryToLane(
     // Sentry already added to lane. Update it.
     return (
       content.slice(0, sentryCLIMatch.index + lane.index) +
-      templates.getFastlaneSnippet(org, project, token).trim() +
+      templates.getFastlaneSnippet(org, project).trim() +
       content.slice(
         sentryCLIMatch.index + sentryCLIMatch[0].length + lane.index,
       )
@@ -96,7 +95,7 @@ function addSentryToLane(
   return (
     content.slice(0, lane.index + lane.length) +
     '\n' +
-    templates.getFastlaneSnippet(org, project, token) +
+    templates.getFastlaneSnippet(org, project) +
     '\n' +
     content.slice(lane.index + lane.length)
   );
@@ -106,7 +105,6 @@ export async function addSentryToFastlane(
   projectPath: string,
   org: string,
   project: string,
-  token: string,
 ): Promise<boolean> {
   const fastFilePath = fastFile(projectPath);
   if (!fastFilePath) {
@@ -133,13 +131,7 @@ export async function addSentryToFastlane(
 
   let newFileContent: string | undefined;
   if (lanes.length === 1) {
-    newFileContent = addSentryToLane(
-      fileContent,
-      lanes[0],
-      org,
-      project,
-      token,
-    );
+    newFileContent = addSentryToLane(fileContent, lanes[0], org, project);
   } else {
     const laneNames = lanes.map((l) => l.name);
     const selectedLane = await askForItemSelection(
@@ -154,7 +146,6 @@ export async function addSentryToFastlane(
       lanes[selectedLane.index],
       org,
       project,
-      token,
     );
   }